+++ package.json
@@ -16,6 +16,7 @@
     "test": "run-p test-node test-react"
   },
   "dependencies": {
+    "aws-sdk": "^2.491.0",
     "axios": "^0.19.0",
     "azure-storage": "^2.10.3",
     "cookie-parser": "^1.4.4",
@@ -27,6 +28,7 @@
     "knex": "^0.16.5",
     "multer": "^1.4.1",
     "multer-azure-storage": "^0.2.0",
+    "multer-s3": "^2.9.0",
     "node-pg-migrate": "^3.21.1",
     "pdftohtmljs": "^0.5.2",
     "prop-types": "^15.7.2",
@@ -73,6 +75,6 @@
   "prettier": {
     "trailingComma": "all",
     "semi": true,
-    "singleQuote": false
+    "singleQuote": true
   }
 }
